requirements_frontend_inference.txt
streamlit>=1.28.2
pandas>=2.1.3
google-cloud-run>=0.10.5
google-cloud-tasks